requirements.txt
pydantic-ai-slim[google]
pydantic-settings
python-telegram-bot[all]
python-dotenv
logfire
sqlalchemy>=2.0.0
aiosqlite>=0.19.0
